+++ conf/start.sh
@@ -3,6 +3,6 @@
 # We build the wallet each time we run the docker and it takes a couple of minutes
 npm run build
 
-cp -r /rui-ui/build/dist/* /var/www/
+cp -r /ui/build/dist/* /var/www/
 
 nginx -g "daemon off;"